Moved Onnx tests to integration tests
